IOS-3512 Split files
